+++ electronic-commerce-backend/product-service/src/main/java/com/jzargo/productservice/service/ImageService.java
@@ -9,10 +9,10 @@ import java.util.List;
 
 public interface ImageService {
 
-    void addImages(byte[][] images, Long productId, Long shopId)
+    void addImages(byte[][] images, Long productId, Integer shopId)
             throws IOException, ProductNotFoundException;
 
-    void addAvatar(byte[] image, Long productId, Long shopId)
+    void addAvatar(byte[] image, Long productId, Integer shopId)
             throws IOException, ProductNotFoundException;
 
     List<byte[]> getImages(Long productId)
